pyproject.toml
[project]
name = "uvproj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.73.1",
    "pillow>=12.2.0",
    "psycopg>=3.3.4",
    "python-dotenv>=1.2.2",
]
